Merge updates to master

+++ src/TimingSegments.cpp
@@ -211,7 +211,7 @@ RString SpeedSegment::ToString(int dec) const
 		+ "f=%.0" + IntToString(dec) + "f=%.0"
 		+ IntToString(dec) + "f=%u";
 	return ssprintf(str.c_str(), GetBeat(), GetRatio(),
-					GetDelay(), GetUnit());
+		GetDelay(), static_cast<unsigned int>(GetUnit()));
 }
 
 vector<float> SpeedSegment::GetValues() const
